+++ src/components/Box/RvBox.tsx
@@ -1,10 +1,15 @@
 /* eslint react/jsx-props-no-spreading: off */
 
 import type { FC } from "react";
-import type { RvBoxFCType } from "../../types/components/Div";
+import type { RvBoxFCType, StyledDivType } from "../../types/components/Div";
 import { StyledDiv } from "./style";
 
-const rvBoxProps: RvBoxFCType = {};
+export const rvDefaultProps: StyledDivType = {
+	width: 30,
+	height: 10,
+	borderRadius: 0.5,
+	boxShadow: "md",
+};
 
 const RvBox: FC<RvBoxFCType> = ({ children, ...props }) => {
 	const newProps = { ...props };