+++ src/components/Box/index.ts
@@ -1,5 +1,6 @@
 import RvBox from "./RvBox";
 import RvFlex from "./RvFlex";
+import RvVStack from "./RvVStack";
 
-export { RvBox, RvFlex };
+export { RvBox, RvFlex, RvVStack };
 export * from "./style";