+++ src/components/Card/index.ts
@@ -1,0 +1,4 @@
+import RvCard from "./RvCard";
+import RvCardHeader from "./RvCardHeader";
+
+export { RvCard, RvCardHeader };